+++ projects/novo-elements/src/elements/form/NovoFormControl.ts
@@ -5,6 +5,7 @@ import { EventEmitter } from '@angular/core';
 import { NovoControlConfig } from './FormControls';
 import { IFieldInteractionEvent } from './FormInterfaces';
 import { notify } from '../../utils/notifier/notifier.util';
+import { IMaskOptions } from './Control';
 
 export class NovoFormControl extends FormControl {
   displayValueChanges: EventEmitter<any> = new EventEmitter<any>();
@@ -19,6 +20,7 @@ export class NovoFormControl extends FormControl {
   tooltipPosition: string;
   tooltipSize?: string;
   tooltipPreline?: boolean;
+  removeTooltipArrow?: boolean;
   initialValue: any;
   valueHistory: any[] = [];
   validators: any;
@@ -43,7 +45,11 @@ export class NovoFormControl extends FormControl {
   layoutOptions?: { order?: string; download?: boolean; labelStyle?: string; draggable?: boolean; iconStyle?: string };
   military?: boolean;
   dateFormat?: string;
+  currencyFormat?: string;
+  startDate?: Date | Number;
+  endDate?: Date | Number;
   textMaskEnabled?: boolean;
+  maskOptions: IMaskOptions;
   allowInvalidDate?: boolean;
   tipWell?: {
     tip: string;
@@ -75,6 +81,7 @@ export class NovoFormControl extends FormControl {
     this.tooltipPosition = control.tooltipPosition;
     this.tooltipSize = control.tooltipSize;
     this.tooltipPreline = control.tooltipPreline;
+    this.removeTooltipArrow = control.removeTooltipArrow;
     this.label = control.label;
     this.name = control.name;
     this.required = control.required;
@@ -88,7 +95,12 @@ export class NovoFormControl extends FormControl {
     this.layoutOptions = control.layoutOptions;
     this.military = control.military;
     this.dateFormat = control.dateFormat;
+    this.currencyFormat = control.currencyFormat;
+    this.startDate = control.startDate;
+    this.endDate = control.endDate;
     this.textMaskEnabled = control.textMaskEnabled;
+    this.textMaskEnabled = control.textMaskEnabled;
+    this.maskOptions = control.maskOptions;
     this.allowInvalidDate = control.allowInvalidDate;
     this.maxlength = control.maxlength;
     this.minlength = control.minlength;