+++ projects/novo-elements/src/elements/form/controls/BaseControl.ts
@@ -48,6 +48,7 @@ export interface NovoControlConfig {
   closeOnSelect?: boolean;
   interactions?: Array<Object>;
   dataSpecialization?: string;
+  dataType?: string;
   appendToBody?: boolean; // Deprecated
   parentScrollSelector?: string;
   description?: string;
@@ -55,6 +56,7 @@ export interface NovoControlConfig {
   tooltipPosition?: string;
   tooltipSize?: string;
   tooltipPreline?: boolean;
+  removeTooltipArrow?: boolean;
   layoutOptions?: {
     order?: string;
     download?: boolean;
@@ -121,6 +123,7 @@ export class BaseControl {
   closeOnSelect: boolean;
   interactions: Array<Object>;
   dataSpecialization: string;
+  dataType: string;
   appendToBody: boolean; // Deprecated
   parentScrollSelector: string;
   description?: string;
@@ -128,6 +131,7 @@ export class BaseControl {
   tooltipPosition?: string;
   tooltipSize?: string;
   tooltipPreline?: boolean;
+  removeTooltipArrow?: boolean;
   layoutOptions?: { order?: string; download?: boolean; labelStyle?: string; draggable?: boolean; iconStyle?: string };
   template?: any;
   customControlConfig?: any;
@@ -200,6 +204,7 @@ export class BaseControl {
     this.closeOnSelect = !!config.closeOnSelect;
     this.interactions = config.interactions;
     this.dataSpecialization = config.dataSpecialization;
+    this.dataType = config.dataType;
     this.appendToBody = !!config.appendToBody;
     if (this.appendToBody) {
       notify(`'appendToBody' has been deprecated. Please remove this attribute.`);
@@ -211,6 +216,7 @@ export class BaseControl {
       this.tooltipPosition = config.tooltipPosition;
       this.tooltipSize = config.tooltipSize;
       this.tooltipPreline = config.tooltipPreline;
+      this.removeTooltipArrow = config.removeTooltipArrow;
     }
     this.template = config.template;
     this.customControlConfig = config.customControlConfig;