+++ projects/novo-elements/src/elements/loading/Loading.module.ts
@@ -2,11 +2,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 // APP
-import { NovoLoadingElement, NovoSpinnerElement } from './Loading';
+import { NovoLoadingElement, NovoSpinnerElement, NovoIsLoadingDirective, NovoLoadedDirective, NovoSkeletonDirective } from './Loading';
 
 @NgModule({
   imports: [CommonModule],
-  declarations: [NovoLoadingElement, NovoSpinnerElement],
-  exports: [NovoLoadingElement, NovoSpinnerElement],
+  declarations: [NovoLoadingElement, NovoSpinnerElement, NovoIsLoadingDirective, NovoLoadedDirective, NovoSkeletonDirective],
+  exports: [NovoLoadingElement, NovoSpinnerElement, NovoIsLoadingDirective, NovoLoadedDirective, NovoSkeletonDirective],
 })
 export class NovoLoadingModule {}